Modularized runInit
