+++ cspell.json
@@ -13,6 +13,7 @@
     "Bluesky",
     "Bridgy",
     "btos",
+    "bunx",
     "callouts",
     "cfworker",
     "cfworkers",
@@ -38,6 +39,7 @@
     "fedify",
     "fediverse",
     "followable",
+    "giget",
     "Guppe",
     "halfyear",
     "highligher",
