+++ packages/cli/src/init/action/dir.ts
@@ -1,0 +1,5 @@
+import { mkdir } from "node:fs/promises";
+import type { InitCommandData } from "../types.ts";
+
+export const makeDirIfHyd = ({ dir }: InitCommandData) =>
+  mkdir(dir, { recursive: true });
