+++ packages/cli/src/init/action/env.ts
@@ -1,0 +1,14 @@
+import { entries, peek, pipeLazy, tap, toArray, when } from "@fxts/core";
+import { notEmpty } from "../../utils.ts";
+import type { InitCommandIo } from "../types.ts";
+import { noticeConfigEnv, noticeEnvKeyValue } from "./notice.ts";
+
+const recommendConfigEnv: InitCommandIo = pipeLazy(
+  (data) => data.env,
+  entries,
+  toArray<Iterable<[string, string]>>,
+  when(notEmpty, tap<[string, string][], void>(noticeConfigEnv)),
+  peek(noticeEnvKeyValue),
+);
+
+export default recommendConfigEnv;
