+++ packages/cli/src/init/lib.ts
@@ -1,44 +1,27 @@
 import {
-  always,
-  concat,
   entries,
   evolve,
   fromEntries,
-  join,
+  isObject,
   map,
+  negate,
   pipe,
-  toArray,
+  throwIf,
   when,
 } from "@fxts/core";
 import { getLogger } from "@logtape/logtape";
 import * as colors from "@std/fmt/colors";
 import { dirname, join as joinPath } from "@std/path";
-import { curry, flow, toMerged, uniq } from "es-toolkit";
+import { toMerged } from "es-toolkit";
 import { readFileSync } from "node:fs";
-import { mkdir, readdir, readFile, writeFile } from "node:fs/promises";
+import { mkdir, readdir, writeFile } from "node:fs/promises";
 import process from "node:process";
 import metadata from "../../deno.json" with { type: "json" };
-import {
-  formatJson,
-  isNotFoundError,
-  match,
-  merge,
-  notEmpty,
-  type RequiredNotNull,
-  runSubCommand,
-} from "../utils.ts";
-import type { InitCommand } from "./command.ts";
-import biome from "./templates/json/biome.json" with { type: "json" };
+import { isNotFoundError, runSubCommand } from "../utils.ts";
 import kv from "./templates/json/kv.json" with { type: "json" };
 import mq from "./templates/json/mq.json" with { type: "json" };
 import pm from "./templates/json/pm.json" with { type: "json" };
 import rt from "./templates/json/rt.json" with { type: "json" };
-import vscodeSettingsForDeno from "./templates/json/vscode-settings-for-deno.json" with {
-  type: "json",
-};
-import vscodeSettings from "./templates/json/vscode-settings.json" with {
-  type: "json",
-};
 import type {
   KvStores,
   MessageQueues,
@@ -94,28 +77,6 @@ const whenHasLabel = <T extends Record<string, { label: string }>>(desc: T) =>
 export const getInstallUrl = (pm: PackageManager) =>
   packageManagers[pm].installUrl;
 
-export function drawDinosaur() {
-  const d = flow(colors.bgBlue, colors.black);
-  const f = colors.blue;
-  console.error(`\
-${d("             ___   ")}  ${f(" _____        _ _  __")}
-${d("            /'_')  ")}  ${f("|  ___|__  __| (_)/ _|_   _")}
-${d("     .-^^^-/  /    ")}  ${f("| |_ / _ \\/ _` | | |_| | | |")}
-${d("   __/       /     ")}  ${f("|  _|  __/ (_| | |  _| |_| |")}
-${d("  <__.|_|-|_|      ")}  ${f("|_|  \\___|\\__,_|_|_|  \\__, |")}
-${d("                   ")}  ${f("                      |___/")}
-`);
-}
-export const logOptions: (options: RequiredNotNull<InitCommand>) => void = (
-  options,
-) =>
-  logger.debug(
-    "Package manager: {packageManager}; " +
-      "web framework: {webFramework}; key–value store: {kvStore}; " +
-      "message queue: {messageQueue}",
-    options,
-  );
-
 export async function isPackageManagerAvailable(
   pm: PackageManager,
 ): Promise<boolean> {
@@ -158,7 +119,6 @@ Then, try look up an actor from your server:
 }
 
 `;
-export const mergeVscSettings = curry(toMerged)(vscodeSettings);
 
 const getDevCommand = (pm: PackageManager) =>
   colors.bold(
@@ -196,152 +156,15 @@ async function isCommandAvailable(
   }
 }
 
-export const addDeps = <
-  T extends {
-    packageManager: PackageManager;
-    dir: string;
-    dependencies: Record<string, string>;
-    dev?: boolean;
-  },
->(data: T) =>
-  addDependencies({
-    ...data,
-    args: getAddDepsArgs(data),
-  });
-
-const getAddDepsArgs = match(
-  ({ dev = false }) => !dev,
-  always(["add"]),
-  ({ packageManager: pm }) => pm === "bun" || pm === "yarn",
-  always(["add", "--dev"]),
-  always(["add", "--save-dev"]),
-);
-
-const addDependencies = <
-  T extends {
-    packageManager: PackageManager;
-    dir: string;
-    dependencies: Record<string, string>;
-    args: string[];
-  },
->({
-  packageManager,
-  dir,
-  dependencies,
-  args = [],
-}: T) =>
-  pipe(
-    dependencies,
-    entryDeps(packageManager),
-    when(notEmpty, (deps) =>
-      pipe(
-        deps,
-        uniq,
-        concat([packageManager, ...args]),
-        toArray,
-        runDepAdd(dir),
-      )),
-  );
-
-export const entryDeps =
-  (pm: PackageManager) => (dependencies: Record<string, string>) =>
-    pipe(
-      dependencies,
-      entries,
-      map(([name, version]) =>
-        `${getPackageName(pm, name)}@${getPackageVersion(pm, version)}`
-      ),
-      toArray,
-    );
-
-const getPackageName = (pm: PackageManager, name: string) =>
-  pm !== "deno" && name.startsWith("npm:")
-    ? name.substring(4)
-    : pm === "deno" && !name.startsWith("npm:")
-    ? `jsr:${name}`
-    : name;
-const getPackageVersion = (pm: PackageManager, version: string) =>
-  pm !== "deno" && version.includes("+")
-    ? version.substring(0, version.indexOf("+"))
-    : version;
-
-const runDepAdd = (dir: string) => (command: string[]) =>
-  runSubCommand(command, {
-    cwd: dir,
-    stdio: "inherit",
-  }).catch((error) => {
-    logger.error(
-      "The command {command} failed with the error: {error}",
-      { command: command.join(" "), error },
-    );
-    throw new Error("Failed to add dependencies.");
-  });
-
-export async function rewriteFile(
-  path: string,
-  content: string | object,
-): Promise<[string, string]> {
-  const prev = await readFileIfExists(path);
-  const data = typeof content === "object"
-    ? formatJson(mergeJson(prev, content))
-    : appendText(prev, content);
-  return [path, data];
-}
-
-const mergeJson = (prev: string, data: object): string =>
-  pipe(prev ? JSON.parse(prev) : {}, merge(data), formatJson);
-
-const appendText = (prev: string, data: string) =>
-  pipe(data.split("\n"), concat(prev.split("\n")), join("\n"));
-
-async function readFileIfExists(path: string): Promise<string> {
-  try {
-    return await readFile(path, "utf8");
-  } catch (e) {
-    if ((e as NodeJS.ErrnoException).code !== "ENOENT") throw e;
-    return "";
-  }
-}
-
 export async function createFile(path: string, content: string): Promise<void> {
   await mkdir(dirname(path), { recursive: true });
   await writeFile(path, content);
 }
 
-export function displayFile(
-  path: string,
-  content: string,
-  emoji: string = "📄",
-  pathColor: (text: string) => string = colors.green,
-) {
-  console.log(pathColor(`${emoji} ${path}`));
-  console.error(colors.gray("─".repeat(60)));
-  console.log(content);
-  console.error(colors.gray("─".repeat(60)) + "\n");
-}
+const isNotExistsError = (e: unknown) =>
+  isObject(e) && "code" in e && e.code === "ENOENT";
 
-export const rewriters = {
-  biome: {
-    path: joinPath("biome.json"),
-    data: biome,
-  },
-  vscExt: {
-    path: joinPath(".vscode", "extensions.json"),
-    data: { recommendations: ["biomejs.biome"] },
-  },
-  vscSet: {
-    path: joinPath(".vscode", "settings.json"),
-    data: vscodeSettings,
-  },
-  vscSetDeno: {
-    path: joinPath(".vscode", "settings.json"),
-    data: vscodeSettingsForDeno,
-  },
-  vscExtDeno: {
-    path: joinPath(".vscode", "extensions.json"),
-    data: { recommendations: ["denoland.vscode-deno"] },
-  },
-} as const;
+export const throwUnlessNotExists = throwIf(negate(isNotExistsError));
 
 export const isDirectoryEmpty = async (
   path: string,
@@ -350,9 +173,7 @@ export const isDirectoryEmpty = async (
     const files = await readdir(path);
     return files.length === 0;
   } catch (e) {
-    if ((e as NodeJS.ErrnoException).code !== "ENOENT") {
-      throw e;
-    }
+    throwUnlessNotExists(e);
     return true;
   }
 };
@@ -366,6 +187,7 @@ export const getNextInitCommand = (
   "--app",
   "--skip-install",
 ];
+
 const createNextAppCommand = (pm: PackageManager): string[] =>
   pm === "deno"
     ? ["deno", "run", "-A", "npm:create-next-app@latest"]
@@ -374,6 +196,7 @@ const createNextAppCommand = (pm: PackageManager): string[] =>
     : pm === "npm"
     ? ["npx", "create-next-app"]
     : [pm, "dlx", "create-next-app"];
+
 export const getNitroInitCommand = (
   pm: PackageManager,
 ): string[] => [
@@ -382,6 +205,7 @@ export const getNitroInitCommand = (
   "nitro",
   ".",
 ];
+
 const createNitroAppCommand = (pm: PackageManager): string[] =>
   pm === "deno"
     ? ["deno", "run", "-A"]
