+++ packages/cli/src/init/mod.ts
@@ -1,2 +1,2 @@
+export { default as runInit } from "./action/mod.ts";
 export { initCommand } from "./command.ts";
-export { default as runInit } from "./action.ts";
