+++ pnpm-workspace.yaml
@@ -38,7 +38,7 @@ catalog:
   '@nestjs/common': ^11.0.1
   amqplib: ^0.10.8
   elysia: ^1.3.8
-  es-toolkit: ^1.39.5
+  es-toolkit: 1.39.5
   express: ^4.0.0
   h3: ^1.15.0
   hono: ^4.0.0
